modules/connection-microservice/location-consumer/requirements.txt
kafka-python == 2.0.2
SQLAlchemy == 1.4.40